requirements.txt
qdrant-client==1.11.1
langchain==0.2.14
langchain-community==0.2.12
streamlit==1.37.1
langchain-huggingface
python-dotenv==1.0.1
langchain-openai==0.1.23